refactor(app): extract useShellConsent hook
Phase 1 step 4 of the god-file cleanup plan. Move the three Allow/Deny
prompt flows out of App.tsx into src/hooks/useShellConsent.ts:

- agent-driven shell_write (read-write mode confirmation)
- close-running-shell tab confirmation
- on-disk session deletion confirmation

Each flow follows the same shape: a per-flow Map<notifId, resolver>
ref, a `resolve*Consent(id, allowed)` callback for the notification
action route, a `prompt*Confirmation(...)` factory that pushes the
notification and returns a Promise<boolean>, and a `hasPending*` query
predicate so the dismissNotification path can settle pending consents
as Deny.

The trust boundary stays in Rust (`shell_write` re-validates the
share mode); the hook only owns the UX gate. SHELL_WRITE_PROMPT_TTL_MS
moved into the hook (~30s under the bridge ack timeout so a late
Allow click can never inject keystrokes after the caller's promise
has timed out).

App.tsx 5886 -> 5763 lines. tsc clean, eslint clean, 305/305 tests
pass. No behavior change.
